everything passed expect stnstr

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: plopes-c <plopes-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/25 20:25:53 by plopes-c          #+#    #+#             */
-/*   Updated: 2022/11/03 19:11:14 by plopes-c         ###   ########.fr       */
+/*   Updated: 2022/11/04 00:44:20 by plopes-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
